+++ src/aui-toggler/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-1364](https://issues.liferay.com/browse/AUI-1364) Make aui-toggler accessible
 
 ## [2.5.0](https://github.com/liferay/alloy-ui/releases/tag/2.5.0)